♻️ refactor: refactor to db env
