+++ src/config/db.ts
@@ -1,0 +1,16 @@
+/* eslint-disable sort-keys-fix/sort-keys-fix , typescript-sort-keys/interface */
+import { createEnv } from '@t3-oss/env-nextjs';
+import { z } from 'zod';
+
+export const getServerDBConfig = () => {
+  return createEnv({
+    client: {
+      NEXT_PUBLIC_ENABLED_SERVER_SERVICE: z.boolean(),
+    },
+    runtimeEnv: {
+      NEXT_PUBLIC_ENABLED_SERVER_SERVICE: process.env.NEXT_PUBLIC_SERVICE_MODE === 'server',
+    },
+  });
+};
+
+export const serverDBEnv = getServerDBConfig();
