+++ src/layout/GlobalProvider/index.tsx
@@ -2,7 +2,7 @@ import dynamic from 'next/dynamic';
 import { cookies } from 'next/headers';
 import { FC, ReactNode } from 'react';
 
-import { getClientConfig } from '@/config/client';
+import { getDebugConfig } from '@/config/debug';
 import { getServerFeatureFlagsValue } from '@/config/featureFlags';
 import { LOBE_LOCALE_COOKIE } from '@/const/locale';
 import {
@@ -26,7 +26,7 @@ let DebugUI: FC = () => null;
 // refs: https://webpack.js.org/plugins/internal-plugins/#constplugin
 if (process.env.NODE_ENV === 'development') {
   // eslint-disable-next-line unicorn/no-lonely-if
-  if (getClientConfig().DEBUG_MODE) {
+  if (getDebugConfig().DEBUG_MODE) {
     DebugUI = dynamic(() => import('@/features/DebugUI'), { ssr: false }) as FC;
   }
 }
